+++ app/src/main/kotlin/at/bitfire/davdroid/ui/intro/BatteryOptimizationsPage.kt
@@ -68,125 +68,6 @@ class BatteryOptimizationsPage: IntroPage {
         BatteryOptimizationsPageContent()
     }
 
-    @Composable
-    private fun BatteryOptimizationsPageContent(
-        model: Model = viewModel()
-    ) {
-        val ignoreBatteryOptimizationsResultLauncher = rememberLauncherForActivityResult(IgnoreBatteryOptimizationsContract) {
-            model.checkBatteryOptimizations()
-        }
-
-        val hintBatteryOptimizations by model.hintBatteryOptimizations.collectAsStateWithLifecycle(false)
-        val shouldBeExempted = model.shouldBeExempted
-        val isExempted = model.isExempted
-        LaunchedEffect(shouldBeExempted, isExempted) {
-            if (shouldBeExempted && !isExempted)
-                ignoreBatteryOptimizationsResultLauncher.launch(BuildConfig.APPLICATION_ID)
-        }
-
-        val hintAutostartPermission by model.hintAutostartPermission.collectAsStateWithLifecycle(false)
-        BatteryOptimizationsPageContent(
-            dontShowBattery = hintBatteryOptimizations == false,
-            onChangeDontShowBattery = model::updateHintBatteryOptimizations,
-            isExempted = isExempted,
-            shouldBeExempted = shouldBeExempted,
-            onChangeShouldBeExempted = model::updateShouldBeExempted,
-            dontShowAutostart = hintAutostartPermission == false,
-            onChangeDontShowAutostart = model::updateHintAutostartPermission,
-            manufacturerWarning = Model.manufacturerWarning
-        )
-    }
-
-
-    @HiltViewModel
-    class Model @Inject constructor(
-        val context: Application,
-        private val settings: SettingsManager
-    ): ViewModel() {
-
-        companion object {
-
-            /**
-             * Whether the request for whitelisting from battery optimizations shall be shown.
-             * If this setting is true or null/not set, the notice shall be shown. Only if this
-             * setting is false, the notice shall not be shown.
-             */
-            const val HINT_BATTERY_OPTIMIZATIONS = "hint_BatteryOptimizations"
-
-            /**
-             * Whether the autostart permission notice shall be shown. If this setting is true
-             * or null/not set, the notice shall be shown. Only if this setting is false, the notice
-             * shall not be shown.
-             *
-             * Type: Boolean
-             */
-            const val HINT_AUTOSTART_PERMISSION = "hint_AutostartPermissions"
-
-            /**
-             * List of manufacturers which are known to restrict background processes or otherwise
-             * block synchronization.
-             *
-             * See https://www.davx5.com/faq/synchronization-is-not-run-as-expected for why this is evil.
-             * See https://github.com/jaredrummler/AndroidDeviceNames/blob/master/json/ for manufacturer values.
-             */
-            private val evilManufacturers = arrayOf("asus", "huawei", "lenovo", "letv", "meizu", "nokia",
-                    "oneplus", "oppo", "samsung", "sony", "vivo", "wiko", "xiaomi", "zte")
-
-            /**
-             * Whether the device has been produced by an evil manufacturer.
-             *
-             * Always true for debug builds (to test the UI).
-             *
-             * @see evilManufacturers
-             */
-            val manufacturerWarning =
-                    (evilManufacturers.contains(Build.MANUFACTURER.lowercase(Locale.ROOT)) || BuildConfig.DEBUG)
-
-            fun isExempted(context: Context) =
-                context.getSystemService<PowerManager>()!!.isIgnoringBatteryOptimizations(BuildConfig.APPLICATION_ID)
-        }
-
-        var shouldBeExempted by mutableStateOf(true)
-            private set
-        var isExempted by mutableStateOf(false)
-            private set
-
-        val hintBatteryOptimizations = settings.getBooleanFlow(HINT_BATTERY_OPTIMIZATIONS)
-
-        val hintAutostartPermission = settings.getBooleanFlow(HINT_AUTOSTART_PERMISSION)
-
-        init {
-            viewModelScope.launch {
-                broadcastReceiverFlow(context, IntentFilter(PermissionUtils.ACTION_POWER_SAVE_WHITELIST_CHANGED), immediate = true).collect {
-                    checkBatteryOptimizations()
-                }
-            }
-        }
-
-        fun checkBatteryOptimizations() {
-            val exempted = isExempted(context)
-            isExempted = exempted
-            shouldBeExempted = exempted
-
-            // if DAVx5 is whitelisted, always show a reminder as soon as it's not whitelisted anymore
-            if (exempted)
-                settings.remove(HINT_BATTERY_OPTIMIZATIONS)
-        }
-
-        fun updateShouldBeExempted(value: Boolean) {
-            shouldBeExempted = value
-        }
-
-        fun updateHintBatteryOptimizations(value: Boolean) {
-            settings.putBoolean(HINT_BATTERY_OPTIMIZATIONS, value)
-        }
-
-        fun updateHintAutostartPermission(value: Boolean) {
-            settings.putBoolean(HINT_AUTOSTART_PERMISSION, value)
-        }
-
-    }
-
 
     @SuppressLint("BatteryLife")
     object IgnoreBatteryOptimizationsContract: ActivityResultContract<String, Unit?>() {
